Moving uniffi_bindgen_kotlin to it's own crate
Moved the code to it's own crate and changed it to use
the `BindingGenerator` trait.
