+++ uniffi_bindgen/askama.toml
@@ -1,9 +1,6 @@
 [general]
 # Directories to search for templates, relative to the crate root.
-dirs = [ "src/scaffolding/templates", "src/bindings/kotlin/templates", "src/bindings/python/templates", "src/bindings/swift/templates", "src/bindings/ruby/templates" ]
-
-[[syntax]]
-name = "kt"
+dirs = [ "src/scaffolding/templates", "src/bindings/python/templates", "src/bindings/swift/templates", "src/bindings/ruby/templates" ]
 
 [[syntax]]
 name = "py"
