+++ uniffi_bindgen_kotlin/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "uniffi_bindgen_kotlin"
+version = "0.1.0"
+edition = "2021"
+
+[[bin]]
+name = "uniffi-bindgen-kotlin"
+path = "src/main.rs"
+
+[dependencies]
+anyhow = "1"
+askama = { version = "0.11", default-features = false, features = ["config"] }
+clap = { version = "2", default-features = false }
+heck = "0.3"
+uniffi_bindgen = { path = "../uniffi_bindgen" }
+paste = "1.0"
+serde = "1"
+toml = "0.5"
+
+[dev-dependencies]
+glob = "0.3"
+uniffi_testing = { path = "../uniffi_testing" }
+uniffi-example-arithmetic = { path = "../examples/arithmetic" }
+uniffi-example-geometry = { path = "../examples/geometry" }
+uniffi-example-rondpoint = { path = "../examples/rondpoint" }
+uniffi-example-todolist = { path = "../examples/todolist" }
+uniffi-example-sprites = { path = "../examples/sprites" }
+uniffi-example-custom-types = { path = "../examples/custom-types" }
+uniffi-fixture-callbacks = { path = "../fixtures/callbacks" }
+uniffi-fixture-ext-types = { path = "../fixtures/ext-types/lib" }
+uniffi-fixture-coverall = { path = "../fixtures/coverall" }
+uniffi-fixture-time = { path = "../fixtures/uniffi-fixture-time" }
