+++ uniffi_bindgen_kotlin/askama.toml
@@ -1,0 +1,8 @@
+[general]
+# Directories to search for templates, relative to the crate root.
+dirs = [ "src/templates" ]
+
+[[syntax]]
+name = "kt"
+
+
