+++ uniffi_bindgen_kotlin/src/gen_kotlin/compounds.rs
@@ -3,7 +3,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 use super::filters;
-use crate::backend::{CodeOracle, CodeType, Literal, TypeIdentifier};
+use uniffi_bindgen::backend::{CodeOracle, CodeType, Literal, TypeIdentifier};
 use askama::Template;
 use paste::paste;
 use std::borrow::Borrow;
