+++ uniffi_bindgen_kotlin/src/gen_kotlin/error.rs
@@ -2,8 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use crate::backend::{CodeDeclaration, CodeOracle, CodeType, Literal, TypeIdentifier};
-use crate::interface::{ComponentInterface, Error};
+use uniffi_bindgen::backend::{CodeDeclaration, CodeOracle, CodeType, Literal, TypeIdentifier};
+use uniffi_bindgen::interface::{ComponentInterface, Error};
 use askama::Template;
 use std::borrow::Borrow;
