+++ uniffi_bindgen_kotlin/src/gen_kotlin/object.rs
@@ -2,22 +2,23 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use crate::backend::{CodeDeclaration, CodeOracle, CodeType, Literal};
-use crate::interface::{CallbackInterface, ComponentInterface};
+use uniffi_bindgen::backend::{CodeDeclaration, CodeOracle, CodeType, Literal};
+use uniffi_bindgen::interface::{ComponentInterface, Object};
 use askama::Template;
 
+// Filters is used by ObjectTemplate.kt, which looks for the filters module here.
 use super::filters;
-pub struct CallbackInterfaceCodeType {
+pub struct ObjectCodeType {
     id: String,
 }
 
-impl CallbackInterfaceCodeType {
+impl ObjectCodeType {
     pub fn new(id: String) -> Self {
         Self { id }
     }
 }
 
-impl CodeType for CallbackInterfaceCodeType {
+impl CodeType for ObjectCodeType {
     fn type_label(&self, oracle: &dyn CodeOracle) -> String {
         oracle.class_name(&self.id)
     }
@@ -32,36 +33,28 @@ impl CodeType for CallbackInterfaceCodeType {
 
     fn helper_code(&self, oracle: &dyn CodeOracle) -> Option<String> {
         Some(format!(
-            "// Helper code for {} callback interface is found in CallbackInterfaceTemplate.kt",
+            "// Helper code for {} class is found in ObjectTemplate.kt",
             self.type_label(oracle)
         ))
     }
 }
 
 #[derive(Template)]
-#[template(syntax = "kt", escape = "none", path = "CallbackInterfaceTemplate.kt")]
-pub struct KotlinCallbackInterface {
-    inner: CallbackInterface,
+#[template(syntax = "kt", escape = "none", path = "ObjectTemplate.kt")]
+pub struct KotlinObject {
+    inner: Object,
 }
 
-impl KotlinCallbackInterface {
-    pub fn new(inner: CallbackInterface, _ci: &ComponentInterface) -> Self {
+impl KotlinObject {
+    pub fn new(inner: Object, _ci: &ComponentInterface) -> Self {
         Self { inner }
     }
-    pub fn inner(&self) -> &CallbackInterface {
+    pub fn inner(&self) -> &Object {
         &self.inner
     }
 }
 
-impl CodeDeclaration for KotlinCallbackInterface {
-    fn initialization_code(&self, oracle: &dyn CodeOracle) -> Option<String> {
-        let code_type = CallbackInterfaceCodeType::new(self.inner.name().into());
-        Some(format!(
-            "{}.register(lib)",
-            code_type.ffi_converter_name(oracle)
-        ))
-    }
-
+impl CodeDeclaration for KotlinObject {
     fn definition_code(&self, _oracle: &dyn CodeOracle) -> Option<String> {
         Some(self.render().unwrap())
     }
@@ -70,8 +63,7 @@ impl CodeDeclaration for KotlinCallbackInterface {
         Some(
             vec![
                 "java.util.concurrent.atomic.AtomicLong",
-                "java.util.concurrent.locks.ReentrantLock",
-                "kotlin.concurrent.withLock",
+                "java.util.concurrent.atomic.AtomicBoolean",
             ]
             .into_iter()
             .map(|s| s.into())
@@ -81,25 +73,25 @@ impl CodeDeclaration for KotlinCallbackInterface {
 }
 
 #[derive(Template)]
-#[template(syntax = "kt", escape = "none", path = "CallbackInterfaceRuntime.kt")]
-pub struct KotlinCallbackInterfaceRuntime {
+#[template(syntax = "kt", escape = "none", path = "ObjectRuntime.kt")]
+pub struct KotlinObjectRuntime {
     is_needed: bool,
 }
 
-impl KotlinCallbackInterfaceRuntime {
+impl KotlinObjectRuntime {
     pub fn new(ci: &ComponentInterface) -> Self {
         Self {
-            is_needed: !ci.iter_callback_interface_definitions().is_empty(),
+            is_needed: !ci.iter_object_definitions().is_empty(),
         }
     }
 }
 
-impl CodeDeclaration for KotlinCallbackInterfaceRuntime {
+impl CodeDeclaration for KotlinObjectRuntime {
     fn definition_code(&self, _oracle: &dyn CodeOracle) -> Option<String> {
-        if !self.is_needed {
-            None
-        } else {
+        if self.is_needed {
             Some(self.render().unwrap())
+        } else {
+            None
         }
     }
 }
