+++ uniffi_bindgen_kotlin/src/gen_kotlin/primitives.rs
@@ -2,8 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use crate::backend::{CodeOracle, CodeType, Literal};
-use crate::interface::{types::Type, Radix};
+use uniffi_bindgen::backend::{CodeOracle, CodeType, Literal};
+use uniffi_bindgen::interface::{types::Type, Radix};
 use askama::Template;
 use paste::paste;
