+++ uniffi_bindgen_kotlin/src/gen_kotlin/record.rs
@@ -2,22 +2,22 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use crate::backend::{CodeDeclaration, CodeOracle, CodeType, Literal};
-use crate::interface::{ComponentInterface, Enum};
+use uniffi_bindgen::backend::{CodeDeclaration, CodeOracle, CodeType, Literal};
+use uniffi_bindgen::interface::{ComponentInterface, Record};
 use askama::Template;
 
 use super::filters;
-pub struct EnumCodeType {
+pub struct RecordCodeType {
     id: String,
 }
 
-impl EnumCodeType {
+impl RecordCodeType {
     pub fn new(id: String) -> Self {
         Self { id }
     }
 }
 
-impl CodeType for EnumCodeType {
+impl CodeType for RecordCodeType {
     fn type_label(&self, oracle: &dyn CodeOracle) -> String {
         oracle.class_name(&self.id)
     }
@@ -26,41 +26,33 @@ impl CodeType for EnumCodeType {
         format!("Type{}", self.id)
     }
 
-    fn literal(&self, oracle: &dyn CodeOracle, literal: &Literal) -> String {
-        if let Literal::Enum(v, _) = literal {
-            format!(
-                "{}.{}",
-                self.type_label(oracle),
-                oracle.enum_variant_name(v)
-            )
-        } else {
-            unreachable!();
-        }
+    fn literal(&self, _oracle: &dyn CodeOracle, _literal: &Literal) -> String {
+        unreachable!();
     }
 
     fn helper_code(&self, oracle: &dyn CodeOracle) -> Option<String> {
         Some(format!(
-            "// Helper code for {} enum is found in EnumTemplate.kt",
+            "// Helper code for {} record is found in RecordTemplate.kt",
             self.type_label(oracle)
         ))
     }
 }
 
 #[derive(Template)]
-#[template(syntax = "kt", escape = "none", path = "EnumTemplate.kt")]
-pub struct KotlinEnum {
-    inner: Enum,
+#[template(syntax = "kt", escape = "none", path = "RecordTemplate.kt")]
+pub struct KotlinRecord {
+    inner: Record,
     contains_object_references: bool,
 }
 
-impl KotlinEnum {
-    pub fn new(inner: Enum, ci: &ComponentInterface) -> Self {
+impl KotlinRecord {
+    pub fn new(inner: Record, ci: &ComponentInterface) -> Self {
         Self {
             contains_object_references: ci.item_contains_object_references(&inner),
             inner,
         }
     }
-    pub fn inner(&self) -> &Enum {
+    pub fn inner(&self) -> &Record {
         &self.inner
     }
     pub fn contains_object_references(&self) -> bool {
@@ -68,7 +60,7 @@ impl KotlinEnum {
     }
 }
 
-impl CodeDeclaration for KotlinEnum {
+impl CodeDeclaration for KotlinRecord {
     fn definition_code(&self, _oracle: &dyn CodeOracle) -> Option<String> {
         Some(self.render().unwrap())
     }
