+++ uniffi_bindgen_kotlin/src/templates/BooleanHelper.kt
