+++ uniffi_bindgen_kotlin/src/templates/CustomTypeTemplate.kt
