+++ uniffi_bindgen_kotlin/src/templates/DurationHelper.kt
