+++ uniffi_bindgen_kotlin/src/templates/EnumTemplate.kt
