+++ uniffi_bindgen_kotlin/src/templates/ErrorTemplate.kt
