+++ uniffi_bindgen_kotlin/src/templates/Float32Helper.kt
