+++ uniffi_bindgen_kotlin/src/templates/Float64Helper.kt
