+++ uniffi_bindgen_kotlin/src/templates/Helpers.kt
