+++ uniffi_bindgen_kotlin/src/templates/Int16Helper.kt
