+++ uniffi_bindgen_kotlin/src/templates/Int32Helper.kt
