+++ uniffi_bindgen_kotlin/src/templates/Int64Helper.kt
