+++ uniffi_bindgen_kotlin/src/templates/Int8Helper.kt
