+++ uniffi_bindgen_kotlin/src/templates/MapTemplate.kt
