+++ uniffi_bindgen_kotlin/src/templates/ObjectRuntime.kt
