+++ uniffi_bindgen_kotlin/src/templates/ObjectTemplate.kt
