+++ uniffi_bindgen_kotlin/src/templates/OptionalTemplate.kt
