+++ uniffi_bindgen_kotlin/src/templates/RecordTemplate.kt
