+++ uniffi_bindgen_kotlin/src/templates/SequenceTemplate.kt
