+++ uniffi_bindgen_kotlin/src/templates/StringHelper.kt
