+++ uniffi_bindgen_kotlin/src/templates/TimestampHelper.kt
