+++ uniffi_bindgen_kotlin/src/templates/TopLevelFunctionTemplate.kt
