+++ uniffi_bindgen_kotlin/src/templates/UInt16Helper.kt
