+++ uniffi_bindgen_kotlin/src/templates/UInt32Helper.kt
