+++ uniffi_bindgen_kotlin/src/templates/UInt64Helper.kt
