+++ uniffi_bindgen_kotlin/src/templates/UInt8Helper.kt
