+++ uniffi_bindgen_kotlin/src/templates/macros.kt
